+++ gojira/ui.go
@@ -19,19 +19,13 @@ func newUi() {
 	app.ui.app = tview.NewApplication()
 	app.ui.pages = tview.NewPages()
 
-	// do I really need those declaration here
+	//FIXME do I really need those declaration here
 	app.ui.calendar = NewCalendar()
 	app.ui.summary = NewSummary()
 	app.ui.dayView = NewDayView()
 	app.ui.errorView = NewErrorView()
 	app.ui.loaderView = NewLoaderView()
 
-	//customModal := func(p tview.Primitive, width, height int) tview.Primitive {
-	//	return tview.NewGrid().
-	//		SetColumns(0, width, 0).
-	//		SetRows(0, height, 0).
-	//		AddItem(p, 1, 1, 1, 1, 0, 0, true)
-	//}
 	app.ui.grid = tview.NewGrid().
 		SetRows(1, 0, 0).
 		SetColumns(0, 0, 27).