+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.5.3] - 2024-03-27
+### Fixed
+- calendar controls not working while focus is on latest issues view
+
+### Changed
+- Set focus on time spent field while adding new worklog
+
 ## [0.5.2] - 2024-03-26
 ### Changed
 - More detailed loader while adding worklogs in a batch
@@ -78,7 +85,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Initial release of gojira
 
-[Unreleased]: https://github.com/jzyinq/gojira/compare/0.5.2...master
+[Unreleased]: https://github.com/jzyinq/gojira/compare/0.5.3...master
+[0.5.3]: https://github.com/jzyinq/gojira/compare/0.5.2...0.5.3
 [0.5.2]: https://github.com/jzyinq/gojira/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/jzyinq/gojira/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/jzyinq/gojira/compare/0.4.0...0.5.0